chore: FavoriteService를 Read와 Write로 분리
